requirements.txt
altgraph==0.17
argon2-cffi==20.1.0
async-generator==1.10
attrs==20.3.0
autopep8==1.5.7
backcall==0.2.0
bleach==3.2.3
cffi==1.14.4
click==7.1.2
colorama==0.4.3
cycler==0.10.0
decorator==4.4.2
defusedxml==0.6.0
distlib==0.3.4
entrypoints==0.3
filelock==3.4.2
Flask==2.0.1
Flask-Login==0.5.0
Flask-SQLAlchemy==2.5.1
future==0.18.2
greenlet==1.1.0
ipykernel==5.4.3
ipython==7.19.0
ipython-genutils==0.2.0
ipywidgets==7.6.3
itsdangerous==2.0.1
jedi==0.18.0
Jinja2==3.0.1
jsonschema==3.2.0
jupyter==1.0.0
jupyter-client==6.1.11
jupyter-console==6.2.0
jupyter-core==4.7.0
jupyterlab-pygments==0.1.2
jupyterlab-widgets==1.0.0
kiwisolver==1.3.1
MarkupSafe==2.0.1
matplotlib==3.4.0
mistune==0.8.4
mpmath==1.2.1
mysql-connector-python==8.0.22
nbclient==0.5.1
nbconvert==6.0.7
nbformat==5.1.2
nest-asyncio==1.5.1
notebook==6.2.0
numpy==1.20.2
packaging==20.8
pandocfilters==1.4.3
parso==0.8.1
pefile==2019.4.18
pickleshare==0.7.5
Pillow==9.0.1
platformdirs==2.4.1
prometheus-client==0.9.0
prompt-toolkit==3.0.14
pycodestyle==2.7.0
pycparser==2.20
pygame==1.9.6
Pygments==2.7.4
pyinstaller==4.0
pyinstaller-hooks-contrib==2020.7
pyparsing==2.4.7
pyrsistent==0.17.3
python-dateutil==2.8.1
pytz==2020.1
pywin32==300
pywin32-ctypes==0.2.0
pywinpty==0.5.7
pyzmq==22.0.2
qtconsole==5.0.2
QtPy==1.9.0
Send2Trash==1.5.0
six==1.15.0
SQLAlchemy==1.4.20
sympy==1.8
terminado==0.9.2
testpath==0.4.4
toml==0.10.2
tornado==6.1
traitlets==5.0.5
virtualenv==20.12.0
wcwidth==0.2.5
webencodings==0.5.1
Werkzeug==2.0.1
widgetsnbextension==3.5.1
